move parse file into import
